Added players. Tile state changes when clicked

+++ cpp/src/tile.h
@@ -1,10 +1,8 @@
 #pragma once
 
-#include <sys/types.h>
-
 class TileState {
   public:
-	static const u_int Empty = 0;
-	static const u_int TakenX = 1;
-	static const u_int TakenO = 2;
+	static const short Empty = 0;
+	static const short TakenX = 1;
+	static const short TakenO = 2;
 };
